Move method for mergeIndex to correct the place
Currently the entry point of function for mergeIndex is in `CommonUtil`
which is not proper. Here in this commit, we will move this to the right
place.
